chore: RoomScreen 에 로직 기능들 Util 로 이동 #21
